requirements.txt
boto3==1.9.151
botocore==1.12.151
certifi==2019.3.9
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.0.3
gunicorn==19.4.5
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
python-dateutil==2.8.0
requests==2.22.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.3
Werkzeug==0.15.4
